requirements.txt
# Core async
aiohttp>=3.9.0
aiofiles>=23.2.0
asyncio-throttle>=1.0.2

# CLI & Config
typer[all]>=0.9.0
rich>=13.7.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
PyYAML>=6.0.1

# Git operations
GitPython>=3.1.40

# Pattern matching
regex>=2023.12.25

# Cryptography
cryptography>=41.0.0

# AWS validation
boto3>=1.34.0
aioboto3>=12.0.0

# File type detection
python-magic-bin>=0.4.14

# Archive handling
py7zr>=0.20.0

# HTTP
aiohttp-socks>=0.8.4

# Utilities
python-dateutil>=2.8.2
humanize>=4.9.0
